Only save valid phone numbers to the database
Saving an invalid phone number resulted in the value `+NoneNone` being
saved to the database. This is an invalid record.

Assume that value is either empty (`None` or `""`) or a `PhoneNumber`.
The field is always accessed through its descriptor, which always return
a `PhoneNumber` when the value's not empty.

Avoids extra work of coercing the value to a `PhoneNumber`.

Reviewed-by: Andrew Bender